Move device extension to utils
