+++ berkeley-mobile/Calendar/CalendarViewController.swift
@@ -228,15 +228,3 @@ extension CalendarViewController {
     }
 }
 
-extension UIDevice {
-    var hasNotch: Bool {
-        if #available(iOS 11.0, *) {
-            if UIApplication.shared.windows.count == 0 { return false }
-            let top = UIApplication.shared.windows[0].safeAreaInsets.top
-            return top > 20
-        } else {
-            // Fallback on earlier versions
-            return false
-        }
-    }
-}
